Small README fix
Hey, thanks for a nice package

I have found that you were using `UserOptions` msg in the README which is not defined. It doesn't made sense to me, so if I'm understanding the code correctly, it should be like this.

Also, maybe it would be nice to know where the `XYZFetched` messages get from when you use MultiWaitable with only `Msg User Options Bookmarks` but there is no mention of those `Msg` with `Fetched` suffix anywhere.